refactor(api): create a new interceptAndRewriteRequest method
